requirements.txt
Django==3.2.3
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-filter==21.1
django-cors-headers==3.12.0
drf-yasg==1.20.0
